Fix: gtag에서 발생하던 버그 수정
